pyproject.toml
[project]
name = "smsssite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3",
    "django",
    "django-admin-logs",
    "django-ckeditor",
    "django-cleanup",
    "django-jazzmin",
    "django-js-asset",
    "django-storages",
    "django-ninja",
    "django-ninja-extra",
    "django-vite",
    "redis",
    "django-redis",
    "gunicorn",
    "Pillow",
    "psycopg2",
    "psycopg2-binary",
    "PyPDF2",
    "requests",
    "s3transfer",
    "xhtml2pdf",
    "python-decouple>=3.8",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "openpyxl>=3.1.5",
]
